thêm email service

+++ accounts/views.py
@@ -12,6 +12,9 @@ from django.contrib.auth import get_user_model
 from rest_framework.permissions import IsAdminUser, IsAuthenticated
 from cart.models import Cart
 from django.contrib.auth import authenticate
+from utils.email import send_verification_email, send_reset_password_email, send_locked_email
+from email_validator import validate_email, EmailNotValidError
+from accounts.models import EmailOTP
 
 User = get_user_model()
 
@@ -26,6 +29,7 @@ class RegisterAPI(APIView):
             return Response({
                 "error": "Email already exists."
             }, status=status.HTTP_400_BAD_REQUEST)  
+            
 
         if not username or not password or not email:
             return Response({'error': 'Username, password, and email are required.'}, status=status.HTTP_400_BAD_REQUEST)
@@ -33,11 +37,54 @@ class RegisterAPI(APIView):
         if User.objects.filter(username=username).exists():
             return Response({'error': 'Username already exists.'}, status=status.HTTP_400_BAD_REQUEST)
 
-        user = User.objects.create_user(username=username, password=password, email=email)
-        user.save()
+        try:
+            validate_email(email)
+        except EmailNotValidError as e:
+            return Response(
+                {"error": str(e)},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+
+        user = User.objects.create_user(username=username, password=password, email=email, is_active=False)
+        send_verification_email(request, user)
 
         return Response({'message': 'User registered successfully.'}, status=status.HTTP_201_CREATED)
-     
+    
+class VerifyAccountAPI(APIView):
+    def post(self, request):
+        email = request.data.get('email')
+        otp = request.data.get('otp')
+        
+        if not email or not otp:
+            return Response({'error': 'Email and OTP are required.'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        if not otp.isdigit() or len(otp) != 6:
+            return Response({'error': 'OTP must be a 6-digit number.'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        try:
+            user = User.objects.get(email=email)
+            otp_res = EmailOTP.objects.get(user=user)
+        except (User.DoesNotExist, EmailOTP.DoesNotExist):
+            return Response({'error': 'Invalid email or OTP.'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        if user.is_active:
+            return Response({'message': 'Account is already verified.'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        if otp_res.is_expired():
+            return Response(
+                {'error': 'OTP has expired.'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+            
+        if otp_res.otp != otp:
+            return Response({'error': 'Invalid OTP.'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        user.is_active = True
+        user.save()
+        otp_res.delete()
+        
+        return Response({'message': 'Account verified successfully.'}, status=status.HTTP_200_OK)
+    
 class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
     def validate(self, attrs):
         username = attrs.get("username")
@@ -177,6 +224,59 @@ class ChangePasswordView(APIView):
         return Response({"message": "Password changed successfully."},
                         status=status.HTTP_200_OK)
 
+class ForgotPasswordView(APIView):
+    def post(self, request):
+        email = request.data.get('email')
+        
+        if not email:
+            return Response({'error': 'Email is required.'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        try:
+            user = User.objects.get(email=email)
+        except User.DoesNotExist:
+            return Response({'error': 'User with this email does not exist.'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        send_reset_password_email(request, user)
+        
+        return Response({'message': 'Password reset OTP sent to your email.'}, status=status.HTTP_200_OK)
+    
+class VerifyPasswordOTPView(APIView):
+    def post(self, request):
+        
+        email = request.data.get("email")
+        otp = request.data.get("otp")
+        new_password = request.data.get("new_password")
+
+        if not email or not otp or not new_password:
+            return Response({"error": "Email, OTP and new_password are required"}, status=status.HTTP_400_BAD_REQUEST)
+        
+        try:
+            user = User.objects.get(email=email)
+            otp_obj = EmailOTP.objects.get(user=user)
+        except (User.DoesNotExist, EmailOTP.DoesNotExist):
+            return Response({'error': 'Invalid email or OTP.'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        if user.check_password(new_password):
+            return Response({"error": "This new password is the same with current password"}, status=status.HTTP_400_BAD_REQUEST)
+
+        if not otp.isdigit() or len(otp) != 6:
+            return Response({"error": "OTP must be a 6-digit number"}, status=status.HTTP_400_BAD_REQUEST)
+
+        if len(new_password) < 6:
+            return Response({"error": "Password must be at least 6 characters"}, status=status.HTTP_400_BAD_REQUEST)
+
+        if otp_obj.is_expired():
+            return Response({"error": "OTP has expired"}, status=status.HTTP_400_BAD_REQUEST)
+
+        if otp_obj.otp != otp:
+            return Response({"error": "Invalid OTP"}, status=status.HTTP_400_BAD_REQUEST)
+
+        user.set_password(new_password)
+        user.save()
+        otp_obj.delete()
+
+        return Response({"message": "Password reset successfully"}, status=status.HTTP_200_OK)
+
 ####ADMIN VIEWS####  
 class UserListAPIView(APIView):
     permission_classes = [IsAdminUser]
@@ -187,13 +287,40 @@ class UserListAPIView(APIView):
     
 class ChangeAccountStatusAPIView(APIView):
     permission_classes = [IsAdminUser]
-    
+
     def patch(self, request, user_id):
         try:
             user = User.objects.get(id=user_id)
-            user.is_active = not user.is_active
-            Cart.objects.filter(user=user).delete()
-            user.save()
-            return Response({'message': 'User status changed successfully.'}, status=status.HTTP_200_OK)
         except User.DoesNotExist:
-            return Response({'error': 'User not found.'}, status=status.HTTP_404_NOT_FOUND)
+            return Response(
+                {"error": "User not found"},
+                status=status.HTTP_404_NOT_FOUND
+            )
+
+        if request.user.id == user.id:
+            return Response(
+                {"error": "You cannot lock your own account"},
+                status=status.HTTP_403_FORBIDDEN
+            )
+
+        if not user.is_active:
+            return Response(
+                {"error": "This account has already been locked and cannot be reactivated"},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+
+        user.is_active = False
+        user.save(update_fields=["is_active"])
+
+        Cart.objects.filter(user=user).delete()
+        
+        send_locked_email(request, user)
+
+        return Response(
+            {
+                "message": "User account has been permanently locked",
+                "user_id": user.id,
+                "is_active": user.is_active
+            },
+            status=status.HTTP_200_OK
+        )
